+++ test/unit/replication.test.ts
@@ -1145,6 +1145,141 @@ describe('replication.test.ts', () => {
         });
     });
     describeParallel('issues', () => {
+        it('#7187 real-time query ignoring the latest changes after deleting and purging data', async () => {
+            const batches = [
+                [
+                    { id: 'foobar', firstName: 'name1' },
+                    { id: 'foobar2', firstName: 'name2' }
+                ]
+            ];
+
+            // create a schema
+            const mySchema = {
+                version: 0,
+                primaryKey: 'id',
+                type: 'object',
+                properties: {
+                    id: {
+                        type: 'string',
+                        maxLength: 100
+                    },
+                    firstName: {
+                        type: 'string'
+                    }
+                }
+            };
+
+            /**
+             * Always generate a random database-name
+             * to ensure that different test runs do not affect each other.
+             */
+            const name = randomToken(10);
+
+            // create a database
+            let db = await createDatabase();
+
+            // Start replication, wait until it's done
+            const firstRep = await createReplication();
+            await firstRep.awaitInitialReplication();
+
+            // Close the database and recreate it
+            await db.close();
+            db = await createDatabase();
+
+            await db.mycollection.find().remove();
+            await db.mycollection.cleanup(0); // If we comment out this line, the test will pass
+
+            const secondRep = await createReplication(false);
+            await secondRep.start();
+            await secondRep.remove();
+
+            // Re-create replication and wait until it's done
+            let items: any[] = [];
+            const itemsQuery = await db.mycollection.find();
+            itemsQuery.$.subscribe(docs => {
+                const mutableDocs = docs.map(doc => doc.toMutableJSON());
+                items = mutableDocs;
+            });
+
+            const thirdRep = await createReplication();
+            await thirdRep.awaitInSync();
+
+            // Add a new batch that represents a document update in the database
+            batches.push([
+                { id: 'foobar', firstName: 'MODIFIED' },
+            ]);
+
+            // Resync and wait until it's done
+            await thirdRep.reSync();
+            await thirdRep.awaitInSync();
+
+            const newQueryResult = await db.mycollection.find({ selector: { id: { $ne: randomToken(10) } } }).exec();
+            assert.deepStrictEqual(
+                newQueryResult.map(d => d.toJSON()),
+                [
+                    {
+                        "id": "foobar",
+                        "firstName": "MODIFIED"
+                    },
+                    {
+                        "id": "foobar2",
+                        "firstName": "name2"
+                    }
+                ],
+                'uncached query result must know about MODIFIED'
+            );
+
+            // THIS FAILS !!!
+            assert.strictEqual(items.find(item => item.id === 'foobar').firstName, 'MODIFIED', 'should have found the modified item');
+            assert.strictEqual(items.length, 2);
+
+            // clean up afterwards
+            db.close();
+
+            function createReplication(autoStart = true) {
+                const replicationState = replicateRxCollection<any, { index: number; }>({
+                    replicationIdentifier: name + 'test-replication',
+                    collection: db.mycollection,
+                    autoStart,
+                    pull: {
+                        batchSize: 3,
+                        handler: async (checkpoint, batchSize) => {
+                            const index = checkpoint?.index ?? 0;
+                            const batchDocs = batches[index];
+                            return {
+                                documents: batchDocs || [],
+                                checkpoint: batchDocs ? { index: index + 1 } : checkpoint,
+                            };
+                        }
+                    },
+                });
+                ensureReplicationHasNoErrors(replicationState);
+                return replicationState;
+            }
+
+            async function createDatabase() {
+                const database = await createRxDatabase({
+                    name,
+                    /**
+                     * By calling config.storage.getStorage(),
+                     * we can ensure that all variations of RxStorage are tested in the CI.
+                     */
+                    storage: config.storage.getStorage(),
+                    cleanupPolicy: {
+                        minimumDeletedTime: 0,
+                    },
+                    localDocuments: true,
+                });
+
+                // create a collection
+                await database.addCollections({
+                    mycollection: {
+                        schema: mySchema
+                    }
+                });
+                return database;
+            }
+        });
         it('upstreamInitialSync() running on all data instead of continuing from checkpoint', async () => {
             const { localCollection, remoteCollection } = await getTestCollections({
                 local: 0,